+++ src/main/java/ru/yandex/money/gradle/plugins/library/dependencies/CheckDependenciesPlugin.java
@@ -12,7 +12,7 @@ import ru.yandex.money.gradle.plugins.library.dependencies.showdependencies.Prin
 import ru.yandex.money.gradle.plugins.library.dependencies.showdependencies.PrintActualOuterDependenciesVersionsTask;
 import ru.yandex.money.gradle.plugins.library.dependencies.showdependencies.PrintInnerDependenciesVersionsTask;
 import ru.yandex.money.gradle.plugins.library.dependencies.showdependencies.PrintOuterDependenciesVersionsTask;
-import ru.yandex.money.gradle.plugins.library.dependencies.snapshot.CheckSnapshotDependenciesTask;
+import ru.yandex.money.gradle.plugins.library.dependencies.snapshot.CheckSnapshotsDependenciesTask;
 
 import javax.annotation.Nonnull;
 import java.util.Set;
@@ -63,7 +63,7 @@ public class CheckDependenciesPlugin implements Plugin<Project> {
     private static final String PRINT_OUTER_DEPENDENCIES_TASK_NAME = "printNewOuterDependenciesVersions";
     private static final String PRINT_ACTUAL_INNER_DEPENDENCIES_TASK_NAME = "printActualInnerDependenciesVersions";
     private static final String PRINT_ACTUAL_OUTER_DEPENDENCIES_TASK_NAME = "printActualOuterDependenciesVersions";
-    private static final String SNAPSHOT_CHECK_TASK_NAME = "checkSnapshotDependencies";
+    private static final String SNAPSHOT_CHECK_TASK_NAME = "checkSnapshotsDependencies";
 
     private static final String PRINT_DEPENDENCIES_TASK_GROUP = "printDependenciesVersions";
 
@@ -195,8 +195,8 @@ public class CheckDependenciesPlugin implements Plugin<Project> {
      * @param project проект
      */
     private static void createCheckSnapshotTask(@Nonnull Project project) {
-        CheckSnapshotDependenciesTask task = project.getTasks()
-                .create(SNAPSHOT_CHECK_TASK_NAME, CheckSnapshotDependenciesTask.class);
+        CheckSnapshotsDependenciesTask task = project.getTasks()
+                .create(SNAPSHOT_CHECK_TASK_NAME, CheckSnapshotsDependenciesTask.class);
 
         task.setGroup(CHECK_DEPENDENCIES_TASK_GROUP);
         task.setDescription("Check snapshot dependecies");