+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Изменила имя таски с "checkSnapshotDependencies" на "checkSnapshotsDependencies" 
 ### NEXT_VERSION_DESCRIPTION_END
 ## [4.4.1]() (06-03-2019)